requirements.txt
os
matplotlib.pyplot
soundfile
datetime
numpy
essentia.standard
json
flask
numpy
wave
io
base64
